Rewrite backend_to_* to not use presence in map as indication that the backend is alive.
With this method, the implementation is greatly simplified, at the (tiny) cost of not freeing keys in the map if there's no more references to the backend. This is not that big of an issue, as currently, there are only 4 backends (if we convert the map to a static array, we can simplify the implementation even more. However, this will require moving the array that defines backends to the utils.hpp header).