Disable sending confirmation back to the client when the server finished writing, as that can lead to a race conflict when reading (as it sends its own messages)
I don't necessarily like this, as it messes up the timings when profiling --- if a write it particularly slow for one write, you will only notice when the *next* write starts, not the current one. Therefore, unless you remember how the program works, you'll likely attribute certain measurements to the wrong call.